Split meal model into meal and ingredients.
